Refactor to multiple modules
