+++ src/core/central_disk_logger/mod.rs
@@ -1,8 +1,11 @@
 pub mod errors;
+pub mod interface;
 pub mod task;
 pub mod traits;
 
-pub use task::{
-    CentralDiskLogger, DiskLoggerMessage, DiskLoggerRegistry, LoggerHandle, LoggerTaskID,
-};
+#[cfg(test)]
+pub mod testing;
+
+pub use interface::{DiskLoggerMessage, DiskLoggerRegistry, LoggerHandle, LoggerTaskID};
+pub use task::CentralDiskLogger;
 pub use traits::MessageLogger;
